About Me content + styling
Added another paragraph to the about me section about being a mathematician who loves to code.

Also styled certain important phrases of the about me section to pop out, such as my job title.

I had the idea at the end to add a tag section to the project cards to display what technologies/tools were used in the project. I have created the files for them but it is not yet implemented.