Gradient styling changes
Fixed problem where gradient would not show up in bg-clip-text elements because the background spanned a larger space than the text did. This occurred in all of the titles, so I restricted the width of the container the text was inside to better fit the space.

Also changed the button on the contact page to have a gradient border with no fill and fill as a gradient when hovered over.